requirements.txt
streamlit==1.37.1
scikit-learn==1.5.1
numpy==2.0.1
google-generativeai>=0.7.2
pytest==8.2.2
